Run gains analysis on per-account basis
This fixes the calculation error that used to occur when non-neutral
investment transactions (= gains) were not all cleanly distributed over
time, which they aren't in practice (see test scenario
simple_savings_5p_with_noise).

For example, if two accounts had yearly gains, one on Jan 1 and the
other on Jan 3, our interpolation between gains would spread out the Jan 1
gain over 363 days the Jan 3 over just the two days in-between. As far
as total as concerned, this is not strictly wrong. But it distorts the
ranged calculations (like per-year and ewm stats).

On real-world data, this change results in slight shifts in the per-year
and ewm analysis results, which is to be expected. The totals (gains,
invested amounts, returns) are completely unaffected.

Note that this change also removes the plotting truncating at
date_of_first_actual_gain because we're not starting the gains DataFrame
at the "beginning of time" anymore, but exactly where each investment
account has its first non-zero balance. This way, there should be no
"settling down" artifacts that would need to be compensated.